refactors and cleans up code
